feat(icp): new inception process for icp and dip
