+++ src/event/event_data/mod.rs
@@ -4,9 +4,15 @@ pub mod interaction;
 pub mod receipt;
 pub mod rotation;
 
-use crate::error::Error;
-use crate::state::{EventSemantics, IdentifierState};
+use crate::{
+    derivation::{self_addressing::SelfAddressing, DerivationCode},
+    error::Error,
+    event_message::serialization_info::{SerializationFormats, SerializationInfo},
+    prefix::IdentifierPrefix,
+    state::{EventSemantics, IdentifierState},
+};
 use serde::{Deserialize, Serialize};
+use serde_hex::{Compact, SerHex};
 
 use self::{
     delegated::{DelegatedInceptionEvent, DelegatedRotationEvent},
@@ -42,5 +48,76 @@ impl EventSemantics for EventData {
             Self::Rct(e) => e.apply_to(state),
             Self::Vrc(e) => e.apply_to(state),
         }
+    }
+}
+
+/// Dummy Event
+///
+/// Used only to encapsulate the prefix derivation process for inception and delegated inception
+#[derive(Serialize, Debug, Clone)]
+pub(super) struct DummyEvent {
+    #[serde(rename = "v")]
+    serialization_info: SerializationInfo,
+    #[serde(rename = "i")]
+    prefix: String,
+    #[serde(rename = "s", with = "SerHex::<Compact>")]
+    sn: u8,
+    #[serde(flatten)]
+    data: EventData,
+}
+
+impl DummyEvent {
+    pub fn derive_inception(
+        icp: InceptionEvent,
+        derivation: SelfAddressing,
+        format: SerializationFormats,
+    ) -> Result<IdentifierPrefix, Error> {
+        Self::derive(EventData::Icp(icp), derivation, format)
+    }
+
+    pub fn derive_delegated_inception(
+        dip: DelegatedInceptionEvent,
+        derivation: SelfAddressing,
+        format: SerializationFormats,
+    ) -> Result<IdentifierPrefix, Error> {
+        Self::derive(EventData::Dip(dip), derivation, format)
+    }
+
+    fn derive(
+        data: EventData,
+        derivation: SelfAddressing,
+        format: SerializationFormats,
+    ) -> Result<IdentifierPrefix, Error> {
+        Ok(IdentifierPrefix::SelfAddressing(
+            derivation.derive(
+                &Self {
+                    serialization_info: SerializationInfo::new(
+                        format,
+                        Self {
+                            serialization_info: SerializationInfo::new(format, 0),
+                            prefix: Self::dummy_prefix(derivation),
+                            sn: 0,
+                            data: data.clone(),
+                        }
+                        .serialize()?
+                        .len(),
+                    ),
+                    prefix: Self::dummy_prefix(derivation),
+                    sn: 0,
+                    data: data,
+                }
+                .serialize()?,
+            ),
+        ))
+    }
+
+    fn serialize(&self) -> Result<Vec<u8>, Error> {
+        self.serialization_info.kind.encode(&self)
+    }
+
+    fn dummy_prefix(derivation: SelfAddressing) -> String {
+        std::iter::repeat("#")
+            .take(derivation.code_len() + derivation.derivative_b64_len())
+            .collect::<String>()
     }
 }
